Revert "client: update window id type (32 -> 64 bits)"
This reverts commit 4cc7feb67b71906d244bfbae396b52fbfe3dc3ce.

+++ client/applist.h
@@ -15,7 +15,7 @@ namespace msgs {
 class App {
     // TODO: extract icon data
 public:
-    typedef quint64 Id;
+    typedef quint32 Id;
     enum {
         INVALID_ID = (Id) -1
     };
